fix Generate_tf.py (EvgenJobTransforms-01-02-08)
        * tagging EvgenJobTransforms-01-02-08
        * fix scripts/EvtGenFromEVNT_tf.py

2014-01-07 E. M. Lobodzinska <ewelina@mail.desy.de>
        * tagging EvgenJobTransforms-01-02-06
        * export to DATAPATH common, share, susycontrol and gencontrol subdirs

2015-01-06 Jim Henderson <James.Henderson@cern.ch>
	* Tagging EvgenJobTransforms-01-02-05
	* Adding share/skeleton.EvtGenFromEVNT.py and scripts/EvtGenFromEVNT_tf.py
	* Add outputTXTFile optional runtime arg for LHEF file output

2014-12-19 E. M. Lobodzinska <ewelina@mail.desy.de>
        * tagging EvgenJobTransforms-01-02-04
        * put the local InstallArea/jobOptions as first in the JOBOPTSEARCHPATH

2014-12-15  David Hall  <david.hall@cern.ch>

	* Tagging: EvgenJobTransforms-01-02-03

...
(Long ChangeLog diff - truncated)

+++ Generators/EvgenJobTransforms/scripts/Generate_tf.py
@@ -18,7 +18,7 @@ import PyJobTransforms.trfArgClasses as trfArgClasses
 
 ## Prodsys1 hack...
 # TODO: Remove!
-ListOfDefaultPositionalKeys=['--AMI', '--AMITag', '--DBRelease', '--asetup', '--athena', '--athenaopts', '--autoConfiguration', '--beamType', '--checkEventCount', '--command', '--conditionsTag', '--ecmEnergy', '--eventAcceptanceEfficiency', '--evgenJobOpts', '--execOnly', '--firstEvent', '--geometryVersion', '--ignoreErrors', '--ignoreFilters', '--ignorePatterns', '--inputGenConfFile', '--inputGeneratorFile', '--jobConfig', '--maxEvents', '--omitFileValidation', '--outputEVNTFile', '--postExec', '--postInclude', '--preExec', '--preInclude', '--randomSeed', '--reportName', '--runNumber', '--showGraph', '--showPath', '--showSteps', '--skipEvents', '--uploadtoami', '--validation']
+ListOfDefaultPositionalKeys=['--AMI', '--AMITag', '--DBRelease', '--asetup', '--athena', '--athenaopts', '--autoConfiguration', '--beamType', '--checkEventCount', '--command', '--conditionsTag', '--ecmEnergy', '--eventAcceptanceEfficiency', '--evgenJobOpts', '--execOnly', '--firstEvent', '--geometryVersion', '--ignoreErrors', '--ignoreFilters', '--ignorePatterns', '--inputGenConfFile', '--inputGeneratorFile', '--jobConfig', '--maxEvents', '--omitFileValidation', '--outputEVNTFile', '--postExec', '--postInclude', '--preExec', '--preInclude', '--randomSeed', '--reportName', '--runNumber', '--showGraph', '--showPath', '--showSteps', '--skipEvents', '--uploadtoami', '--validation', '--outputTXTFile']
 
 
 class EvgenExecutor(athenaExecutor):
@@ -47,17 +47,26 @@ class EvgenExecutor(athenaExecutor):
                 zf.close()
 
         def mk_jo_proxy(targetbasepath, pkgname, proxypath="_joproxy", addtosearch=True):
-            "Make a JO proxy dir such that the MCxxJobOptions/dddd dirs' contents are found via include(MCxxJobOptions/yyyy)"
-            if os.path.exists(proxypath):
-                shutil.rmtree(proxypath)
-            os.mkdir(proxypath)
+            "Make a JO proxy dir such that the MCxxJobOptions/dddd dirs contents are found via include(MCxxJobOptions/yyyy)"
+            if proxypath:
+                if os.path.exists(proxypath):
+                    shutil.rmtree(proxypath)
+                os.mkdir(proxypath)
+            os.environ['LOCAL_INSTALL_DIR'] = (os.environ['JOBOPTSEARCHPATH']).split(":")[0]    
             for d in ("common", "share", "gencontrol", "susycontrol"):
                 # TODO: we could _maybe_ add the appropriate share/DSIDxxxx/ dir to the path based on the jobConfig arg... too much magic?
                 dpath = os.path.join(proxypath, d)
-                os.mkdir(dpath)
-                os.symlink(os.path.join(targetbasepath, d), os.path.join(dpath, pkgname))
+                if proxypath:
+                    os.mkdir(dpath)
+                    os.symlink(os.path.join(targetbasepath, d), os.path.join(dpath, pkgname))
                 if addtosearch:
-                    os.environ["JOBOPTSEARCHPATH"] = dpath + ":" + os.environ["JOBOPTSEARCHPATH"]
+                    os.environ["JOBOPTSEARCHPATH"] = dpath+":"+os.environ["JOBOPTSEARCHPATH"]
+                    os.environ["DATAPATH"] =os.path.join(targetbasepath, d)+":"+os.environ["DATAPATH"]
+#                    print("EA: DATAPATH=%s;") % os.environ["DATAPATH"]
+#		    print("EA: dpath=%s;") % (dpath)
+#                    print("EA: realpath=%s;") % os.path.realpath(dpath)
+            os.environ["JOBOPTSEARCHPATH"] = os.environ['LOCAL_INSTALL_DIR']+":"+os.environ["JOBOPTSEARCHPATH"]
+
 
         ## Handle locating of evgen job options / fragments, either from a tarball or CVMFS
         if "evgenJobOpts" in self._trf.argdict: ## Use a specified JO tarball
@@ -78,6 +87,7 @@ class EvgenExecutor(athenaExecutor):
                 msg.info('Evgen tarball download success: %s' % output)
             ## Expand tarball
             expand_if_archive(tarball)
+            mk_jo_proxy(os.getcwd(), "MC14JobOptions", "")
             ## Source setup script (requires some shenanigans to update the Python env robustly)
             # TODO: trf framework now bans use of exec()...
             #import subprocess
